feat: migrate MetadataPopover component to v2 REST API
